// src/types/dtos/role.ts
/**
 * Generated by orval v6.1.1 🍺
 * Do not edit manually.
 * Nightborn Dev Exercise API
 * Swagger for Nightborn-dev-exercise-back-end project.

Schemes: [http, https]
 * OpenAPI spec version: 1.0.0
 */

/**
 * A role for an user
 */
export type Role = 'Administrator' | 'RegularUser';


export const Role = {
  Administrator: 'Administrator' as Role,
  RegularUser: 'RegularUser' as Role,
};